+++ src/collective/ploneboard/interfaces.py
@@ -1,6 +1,6 @@
 # -*- coding: UTF-8 -*-
 from plone.app.textfield import RichText
-from plone.directives import form
+from zope.interface import Interface
 from zope.publisher.interfaces.browser import IDefaultBrowserLayer
 
 from collective.ploneboard import _
@@ -10,17 +10,17 @@ class ICollectivePloneboardLayer(IDefaultBrowserLayer):
     """Marker interface that defines a browser layer."""
 
 
-class IMessageboard(form.Schema):
+class IMessageboard(Interface):
     """
     """
 
 
-class ITopic(form.Schema):
+class ITopic(Interface):
     """
     """
 
 
-class IConversation(form.Schema):
+class IConversation(Interface):
     """
     """